Bug in differential / ersky9x

+++ src/open9x.cpp
@@ -1059,7 +1059,7 @@ void checkTHR()
 {
   if (g_model.disableThrottleWarning) return;
 
-  int thrchn = (2-(stickMode&1));//stickMode=0123 -> thr=2121
+  int8_t thrchn = (2-(stickMode&1)); //stickMode=0123 -> thr=2121
 
 #ifdef SIMU
   int16_t lowLim = THRCHK_DEADBAND - 1024 ;
